+++ .pre-commit-config.yaml
@@ -31,6 +31,7 @@ repos:
     rev: 5.13.2
     hooks:
       - id: isort
+        args: ["--profile=black"]
 
   - repo: https://github.com/psf/black-pre-commit-mirror
     rev: 24.4.2
